fix pre-commit
