+++ homework1/task04.py
@@ -6,8 +6,8 @@ Given four lists A, B, C, D of integer values,
 
 We guarantee, that all A, B, C, D have same length of N where 0 ≤ N ≤ 1000.
 """
-from typing import List
 import itertools
+from typing import List
 
 
 def check_sum_of_four_first_solution(
