+++ homework1/task05.py
@@ -9,9 +9,9 @@ Examples:
     nums = [1, 3, -1, -3, 5, 3, 6, 7], k = 3
     result = 16
 """
-from typing import List
-from collections import deque
 import itertools
+from collections import deque
+from typing import List
 
 
 def find_maximal_subarray_sum(nums: List[int], k: int) -> int:
