+++ homework1/tests/test_file.txt
@@ -1,10 +1,10 @@
-[21, 86, 28, 88, 87, 39, 18, 60, 61, 9]
-[87, 80, 73, 81, 4, 44, 98, 64, 87, 48]
-[27, 84, 74, 68, 16, 63, 96, 73, 27, 45]
-[9, 12, 35, 90, 38, 47, 98, 95, 43, 89]
-[74, 49, 53, 89, 29, 20, 73, 22, 91, 57]
-[46, 43, 100, 22, 3, 27, 94, 91, 41, 20]
-[10, 58, 60, 60, 99, 35, 96, 83, 68, 94]
-[38, 2, 80, 66, 28, 89, 9, 44, 41, 37]
-[77, 19, 75, 20, 4, 16, 24, 42, 26, 16]
-[85, 47, 52, 64, 12, 88, 60, 19, 89, 82]
+[71, 64, 57, 26, 18, 73, 81, 46, 9, 9]
+[34, 31, 98, 25, 3, 64, 45, 4, 31, 61]
+[8, 17, 97, 97, 9, 36, 37, 39, 16, 11]
+[92, 6, 18, 88, 9, 88, 35, 44, 63, 55]
+[15, 63, 34, 96, 19, 96, 90, 16, 74, 56]
+[72, 38, 2, 77, 28, 4, 45, 79, 40, 88]
+[62, 8, 99, 60, 84, 8, 71, 30, 72, 2]
+[47, 75, 9, 98, 100, 52, 10, 10, 63, 19]
+[53, 55, 4, 21, 36, 35, 19, 80, 25, 15]
+[32, 93, 44, 15, 86, 94, 44, 9, 5, 54]
